requirements.txt
################################################################################
# This requirements file has been automatically generated from `Pipfile` with
# `pipenv-to-requirements`
#
#
# This has been done to maintain backward compatibility with tools and services
# that do not support `Pipfile` yet.
#
# Do NOT edit it directly, use `pipenv install [-d]` to modify `Pipfile` and
# `Pipfile.lock` and then regenerate `requirements*.txt`.
################################################################################

asgiref==3.2.10 ; python_version >= '3.5'
django-appconf==1.0.4
django-cleanup==5.1.0
django-crispy-forms==1.9.2
django-dotenv==1.4.2
django-filter==2.4.0
django-imagekit==4.0.2
django==3.1.2
djangorestframework-jwt==1.11.0
djangorestframework==3.12.1
mysqlclient==2.0.1
pilkit==2.0
pillow==7.2.0
pyjwt==1.7.1
pytz==2020.1
six==1.15.0 ; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
sqlparse==0.4.1 ; python_version >= '3.5'
